+++ test/cases/escape-render/webpack.config.js
@@ -1,17 +1,20 @@
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const PugPlugin = require('pug-plugin');
 const webRootPath = path.join(__dirname, 'public/');
 
 module.exports = {
   mode: 'production',
-  entry: {},
-  plugins: [
-    new HtmlWebpackPlugin({
-      filename: path.join(webRootPath, '/index.html'),
-      template: './src/index.pug',
-      inject: false,
-    }),
-  ],
+
+  output: {
+    path: webRootPath,
+    publicPath: '',
+  },
+
+  entry: {
+    index: 'src/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
 
   module: {
     rules: [