+++ test/cases/extends-alias/webpack.config.js
@@ -1,24 +1,28 @@
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const PugPlugin = require('pug-plugin');
 const basePath = path.resolve(__dirname);
 const webRootPath = path.join(__dirname, 'public/');
 
 module.exports = {
   mode: 'production',
-  entry: {},
   resolve: {
     alias: {
       Includes: path.join(basePath, 'src/includes/'),
       Template: path.join(basePath, 'src/template/'),
     },
   },
-  plugins: [
-    new HtmlWebpackPlugin({
-      filename: path.join(webRootPath, '/index.html'),
-      template: './src/template/index.pug',
-      inject: false,
-    }),
-  ],
+
+  output: {
+    path: webRootPath,
+    publicPath: '',
+  },
+
+  entry: {
+    index: 'src/template/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
+
   module: {
     rules: [
       {