+++ test/cases/include-relative/webpack.config.js
@@ -1,18 +1,21 @@
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const PugPlugin = require('pug-plugin');
 const webRootPath = path.join(__dirname, 'public/');
-const publicPath = '/assets/';
 
 module.exports = {
   mode: 'production',
-  entry: {},
-  plugins: [
-    new HtmlWebpackPlugin({
-      filename: path.join(webRootPath, '/index.html'),
-      template: './src/template/index.pug',
-      inject: false,
-    }),
-  ],
+
+  output: {
+    path: webRootPath,
+    publicPath: '',
+  },
+
+  entry: {
+    index: 'src/template/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
+
   module: {
     rules: [
       {