+++ test/cases/javascript-inline-loader/webpack.config.js
@@ -3,9 +3,16 @@ const basePath = path.resolve(__dirname);
 
 module.exports = {
   mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '',
+  },
+
   entry: {
     index: './src/index.js',
   },
+
   resolve: {
     alias: {
       Includes: path.join(basePath, 'src/includes/'),
@@ -13,7 +20,9 @@ module.exports = {
       Template: path.join(basePath, 'src/template/'),
     },
   },
+
   plugins: [],
+
   module: {
     // Note: the loader name is defined at `resolveLoader.alias` as 'pug-loader': '@webdiscus/pug-loader'
     // disable a pug-loader in webpack config to use inline pug-loader in require query, e.g.: